Use pickle to send exceptions from the child to the parent
With the old mechanism, there was a linting error trying to assign the
transmitted error code to 'exception.errno'. I suppose the linter didn't
realize that the received exception had an errno member. We can avoid
all of this by using Python's built-in serialization facility rather
than rolling a custom one.