Refactor _make_eof_intr into _get_intr_eof
The former existed in order to defer the generation of these values (to
avoid a costly import of the module), and stored them in global
variables. However, since the function was called when the PtyProcess
was created, it was going to be expensive regardless. The new version
caches its results (using functools.cache rather than globals) and
returns those values on subsequent calls, meaning that the values are
generated only if/when needed.